refactor(web): API クライアントの依存方向を整える
queryClient と adminUserKeys を lib/ 配下に切り出し、 lib/api/client から providers/ や hooks/ への参照を解消する。
