+++ web/src/lib/api/client.ts
@@ -1,13 +1,13 @@
 import * as Sentry from "@sentry/react";
 import createClient from "openapi-fetch";
-import { adminUserKeys } from "~/hooks/useAdminUser";
 import {
   getBearerToken,
   getSessionToken,
   removeAuthToken,
 } from "~/lib/auth-token";
-import { queryClient } from "~/providers/QueryProvider";
+import { queryClient } from "~/lib/query-client";
 import type { paths } from "~/types/api";
+import { adminUserKeys } from "./keys";
 
 class ApiError extends Error {
   status: number;
@@ -21,7 +21,8 @@ class ApiError extends Error {
 const API_BASE = import.meta.env.PUBLIC_API_URL || "";
 
 // 401 が返って admin token が原因と判定できれば、token を破棄してフォールバックで再試行する。
-// 並行リクエストで判定がぶれないよう、送信時の Authorization ヘッダーを起点に判定する。
+// session token は不変なので、「送ったヘッダが session token と一致しない」を起点に判定する。
+// admin token は破棄後も並行リクエストで判定がぶれないよう、現在値ではなく送信ヘッダ側を見る。
 const fetchWithAuthRetry = async (request: Request): Promise<Response> => {
   const retryRequest = request.clone();
   const response = await fetch(request);
@@ -30,8 +31,7 @@ const fetchWithAuthRetry = async (request: Request): Promise<Response> => {
   const sentAuth = retryRequest.headers.get("Authorization");
   if (!sentAuth) return response;
   const sessionToken = getSessionToken();
-  const wasAdminAuth = !sessionToken || sentAuth !== `Bearer ${sessionToken}`;
-  if (!wasAdminAuth) return response;
+  if (sessionToken && sentAuth === `Bearer ${sessionToken}`) return response;
 
   removeAuthToken();
   queryClient.invalidateQueries({ queryKey: adminUserKeys.current });
