+++ web/src/lib/api/keys.ts
@@ -1,0 +1,3 @@
+export const adminUserKeys = {
+  current: ["adminUser", "current"] as const,
+};
